# src/main/java/com/behl/overseer/dto/JokeResponseDto.java
package com.behl.overseer.dto;

import com.fasterxml.jackson.databind.PropertyNamingStrategies;
import com.fasterxml.jackson.databind.annotation.JsonNaming;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.Builder;
import lombok.Getter;
import lombok.extern.jackson.Jacksonized;

@Getter
@Builder
@Jacksonized
@JsonNaming(value = PropertyNamingStrategies.UpperCamelCaseStrategy.class)
@Schema(title = "Joke", accessMode = Schema.AccessMode.READ_ONLY)
public class JokeResponseDto {
	
	// Generated joke returned by the protected joke endpoint.
	private String joke;

}